requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
cryptography==36.0.2
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-import-export==2.8.0
django-rest-framework==0.1.0
django-unicorn==0.44.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
orjson==3.6.7
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
shortuuid==1.0.8
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.2.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0